Many function were divided into several classes and packages. SingleHeroError, ChooseHeroError was modified with image and background color. Loading.kt component was added
