+++ app/src/main/java/com/example/marvel_app_project/assets/HeroesData.kt
@@ -1,7 +1,7 @@
 package com.example.marvel_app_project.assets
 
 import androidx.compose.ui.graphics.Color
-import com.example.marvel_app_project.models.HeroUI
+import com.example.marvel_app_project.models.UiLayer.HeroUI
 
 object SampleData{
     val heroUISamples = listOf(
