+++ app/src/main/java/com/example/marvel_app_project/models/DataLayer/MoshiResponse.kt
@@ -1,4 +1,4 @@
-package com.example.marvel_app_project.models
+package com.example.marvel_app_project.models.DataLayer
 
 data class MoshiResponse(
     val data: MoshiResponseData
