+++ app/src/main/java/com/example/marvel_app_project/models/DataLayer/MoshiResponseData.kt
@@ -1,5 +1,6 @@
-package com.example.marvel_app_project.models
+package com.example.marvel_app_project.models.DataLayer
 
+import com.example.marvel_app_project.models.DataLayer.HeroMoshi
 import com.squareup.moshi.Json
 
 data class MoshiResponseData(
