+++ app/src/main/java/com/example/marvel_app_project/network/Either/Either.kt
@@ -1,4 +1,4 @@
-package com.example.marvel_app_project.network
+package com.example.marvel_app_project.network.Either
 
 sealed class Either<out F, out S> {
