+++ app/src/main/java/com/example/marvel_app_project/network/Either/EitherCallAdapter.kt
@@ -1,6 +1,6 @@
-package com.example.marvel_app_project.network
+package com.example.marvel_app_project.network.Either
 
-import com.example.marvel_app_project.models.ErrorResponse
+import com.example.marvel_app_project.models.DataLayer.ErrorResponse
 import retrofit2.Call
 import retrofit2.CallAdapter
 import java.lang.reflect.Type
