+++ app/src/main/java/com/example/marvel_app_project/network/HeroApiService.kt
@@ -1,71 +1,17 @@
 package com.example.marvel_app_project.network
 
-
-
-import com.example.marvel_app_project.BuildConfig
-import com.example.marvel_app_project.models.ErrorResponse
-import com.example.marvel_app_project.models.MoshiResponse
-import com.example.marvel_app_project.models.MoshiResponseData
+import com.example.marvel_app_project.models.DataLayer.ErrorResponse
+import com.example.marvel_app_project.models.DataLayer.MoshiResponse
+import com.example.marvel_app_project.network.Either.Either
+import com.example.marvel_app_project.network.Either.EitherCallAdapterFactory
 import com.squareup.moshi.Moshi
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.converter.scalars.ScalarsConverterFactory
 import retrofit2.http.GET
 import retrofit2.http.Path
-import java.math.BigInteger
-import java.security.MessageDigest
-import java.sql.Timestamp
-
 
 private const val BASE_URL = "http://gateway.marvel.com/v1/public/"
-
-
-class ParceConstants{
-    companion object{
-        val ts = Timestamp(System.currentTimeMillis()).time.toString()
-        const val limit = "10"
-
-        const val API_KEY = BuildConfig.API_KEY
-        const val PRIVATE_API_KEY = BuildConfig.PRIVATE_API_KEY
-
-        fun hash():String{
-            val inputHashString = "$ts$PRIVATE_API_KEY$API_KEY"
-            val md = MessageDigest.getInstance("MD5")
-            return BigInteger(1, md.digest(inputHashString.toByteArray())).toString(16).padStart(32,'0')
-        }
-    }
-}
-
-val authInterceptor = Interceptor{chain ->
-    val originalRequest = chain.request()
-
-    val newHttpUrl = originalRequest.url.newBuilder()
-        .addQueryParameter("apikey", ParceConstants.API_KEY)
-        .addQueryParameter("ts", ParceConstants.ts)
-        .addQueryParameter("hash", ParceConstants.hash())
-        .addQueryParameter("limit", ParceConstants.limit)
-        .build()
-
-    val newRequest = originalRequest.newBuilder()
-        .url(newHttpUrl)
-        .build()
-
-    chain.proceed(newRequest)
-}
-
-val loggingInterceptor = HttpLoggingInterceptor()
-    .apply {
-        setLevel(HttpLoggingInterceptor.Level.BASIC)
-    }
-
-val client = OkHttpClient.Builder()
-    .addNetworkInterceptor(loggingInterceptor)
-    .addNetworkInterceptor(authInterceptor)
-    .build()
-
 
 val moshi = Moshi.Builder()
     .add(com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory())
@@ -80,16 +26,15 @@ private val retrofit = Retrofit.Builder()
     .baseUrl(BASE_URL)
     .build()
 
-
 interface HeroApiService{
     @GET("characters")
     suspend fun getMarvelCharacters(
-    ):Either<ErrorResponse, MoshiResponse>
+    ): Either<ErrorResponse, MoshiResponse>
 
     @GET("characters/{characterId}")
     suspend fun getSingleMarvelCharacter(
         @Path("characterId") id: Int
-    ):Either<ErrorResponse, MoshiResponse>
+    ): Either<ErrorResponse, MoshiResponse>
 
 }
