+++ app/src/main/java/com/example/marvel_app_project/ui/pages/choosehero/ChooseHeroesUiState.kt
@@ -1,0 +1,9 @@
+package com.example.marvel_app_project.ui.pages.choosehero
+
+import com.example.marvel_app_project.models.UiLayer.HeroUI
+
+sealed interface ChooseHeroesUiState{
+    data class Success(val heroUIValues: List<HeroUI>): ChooseHeroesUiState
+    data class Error(val errorMessage: String, val reserveHeroUiValues: List<HeroUI>): ChooseHeroesUiState
+    object Loading: ChooseHeroesUiState
+}
