+++ app/src/main/java/com/example/marvel_app_project/ui/pages/choosehero/screens/ChooseHeroResult.kt
@@ -1,4 +1,4 @@
-package com.example.marvel_app_project.ui.pages.ChooseHero
+package com.example.marvel_app_project.ui.pages.choosehero.screens
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.ExperimentalFoundationApi
@@ -14,8 +14,6 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -23,45 +21,22 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Path
-import com.example.marvel_app_project.models.HeroUI
+import com.example.marvel_app_project.models.UiLayer.HeroUI
 import com.example.marvel_app_project.ui.components.ChooseHeroHeader
 import com.example.marvel_app_project.ui.components.HeroCard
 import com.example.marvel_app_project.ui.theme.Sizes
 import com.example.marvel_app_project.ui.theme.Spaces
 
-@Composable
-fun ChooseHeroScreen(
-    heroesUiState: ChooseHeroesUiState,
-    onHeroImageTaped:(Int, String) -> Unit) {
-
-
-    when(heroesUiState){
-        is ChooseHeroesUiState.Loading -> ChooseHeroLoading()
-        is ChooseHeroesUiState.Error -> ChooseHeroError(
-            errorMessage = heroesUiState.errorMessage,
-            heroValues = heroesUiState.reserveHeroUiValues,
-            onHeroImageTaped = onHeroImageTaped
-        )
-        is ChooseHeroesUiState.Success -> ChooseHeroResult(
-            heroValues = heroesUiState.heroUIValues,
-            onHeroImageTaped = onHeroImageTaped
-        )
-    }
-
-
-}
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun ChooseHeroResult(
-    //heroViewModel: HeroViewModel = viewModel(),
     heroValues: List<HeroUI>,
     onHeroImageTaped:(Int, String) -> Unit){
 
-    //val heroValues by heroViewModel.reserveHeroUIState.collectAsState()
     val lazyListState = rememberLazyListState()
     val snapBehavior = rememberSnapFlingBehavior(lazyListState = lazyListState)
     val rectangleColorState = remember {
-        mutableStateOf(Color(119, 3,8))
+        mutableStateOf(Color(0xFF770308))
     }
     val latestIndex = remember {
         mutableStateOf(lazyListState.layoutInfo.visibleItemsInfo.lastOrNull()?.index)
@@ -80,8 +55,6 @@ fun ChooseHeroResult(
             }
         }
     }
-
-    //Text(text=heroesValues, color = MaterialTheme.colorScheme.onSecondary)
 
     Box (modifier = Modifier.fillMaxSize()){
         Column (
@@ -139,24 +112,3 @@ fun ChooseHeroResult(
         }
     }
 }
-
-@Composable
-fun ChooseHeroLoading(){
-    Text(text = "loading...")
-}
-
-@Composable
-fun ChooseHeroError(errorMessage: String, heroValues: List<HeroUI>, onHeroImageTaped:(Int, String) -> Unit){
-
-    Text(text = "$errorMessage Can't reach data from server", color = MaterialTheme.colorScheme.onSecondary)
-
-    ChooseHeroResult(heroValues = heroValues, onHeroImageTaped = onHeroImageTaped)
-}
-
-//@Preview(showBackground = true)
-//@Composable
-//fun GreetingPreview() {
-//    Marvel_app_projectTheme {
-//        ChooseHeroResult(heroValues = SampleData.heroUISamples, onHeroImageTaped = {})
-//    }
-//}
