+++ app/src/main/java/com/example/marvel_app_project/ui/pages/singlehero/SingleHeroUiState.kt
@@ -1,0 +1,9 @@
+package com.example.marvel_app_project.ui.pages.singlehero
+
+import com.example.marvel_app_project.models.UiLayer.HeroUI
+
+sealed interface SingleHeroUiState{
+    data class Success(val singleHeroUIValue: HeroUI): SingleHeroUiState
+    data class Error(val errorMessage:String, val reserveSingleHeroUiValue: HeroUI): SingleHeroUiState
+    object Loading: SingleHeroUiState
+}
