+++ app/src/main/java/com/example/marvel_app_project/ui/pages/singlehero/SingleHeroViewModel.kt
@@ -1,30 +1,22 @@
-package com.example.marvel_app_project.ui.pages.SingleHero
+package com.example.marvel_app_project.ui.pages.singlehero
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.marvel_app_project.models.HeroUI
-import com.example.marvel_app_project.models.toSingleUI
-import com.example.marvel_app_project.models.toStringType
-import com.example.marvel_app_project.network.Either
+import com.example.marvel_app_project.models.UiLayer.HeroUI
+import com.example.marvel_app_project.models.DataLayer.toSingleUI
+import com.example.marvel_app_project.models.DataLayer.toStringType
+import com.example.marvel_app_project.network.Either.Either
 import com.example.marvel_app_project.network.HeroApi
-import com.example.marvel_app_project.ui.pages.ChooseHero.ChooseHeroViewModel
+import com.example.marvel_app_project.ui.pages.choosehero.ChooseHeroViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
-sealed interface SingleHeroUiState{
-    data class Success(val singleHeroUIValue: HeroUI): SingleHeroUiState
-    data class Error(val errorMessage:String, val reserveSingleHeroUiValue: HeroUI): SingleHeroUiState
-    object Loading: SingleHeroUiState
-}
 class SingleHeroViewModel: ViewModel() {
 
     private var _reserveSingleHeroUIState = MutableStateFlow(HeroUI())
-    val reserveSingleHeroUIState: StateFlow<HeroUI> = _reserveSingleHeroUIState.asStateFlow()
 
     var singleHeroUIState: SingleHeroUiState by mutableStateOf(SingleHeroUiState.Loading)
 
@@ -44,8 +36,7 @@ class SingleHeroViewModel: ViewModel() {
         }
     }
 }
-
-    fun reserveUpdateHero(heroName: String):HeroUI{
+    fun reserveUpdateHero(heroName: String): HeroUI {
         val chooseHeroViewModel = ChooseHeroViewModel()
         val currentHeroValues = chooseHeroViewModel.reserveHeroUIState.value
         _reserveSingleHeroUIState.value = currentHeroValues.find { it.name == heroName }?: HeroUI()
