+++ app/src/main/java/com/example/marvel_app_project/ui/pages/singlehero/screens/SingleHeroError.kt
@@ -1,0 +1,58 @@
+package com.example.marvel_app_project.ui.pages.singlehero.screens
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import com.example.marvel_app_project.R
+import com.example.marvel_app_project.models.UiLayer.HeroUI
+import com.example.marvel_app_project.ui.theme.Shapes
+import com.example.marvel_app_project.ui.theme.Sizes
+import com.example.marvel_app_project.ui.theme.Spaces
+import com.example.marvel_app_project.ui.theme.interFamily
+
+@Composable
+fun SingleHeroError(errorMessage: String, hero: HeroUI, navigateUp: () -> Unit){
+    Column {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = Spaces.errorColumn)
+                .clip(Shapes.medium)
+                .background(color = MaterialTheme.colorScheme.secondary),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ){
+            Image(
+                modifier = Modifier.size(Sizes.noInternetLogo.size),
+                painter = painterResource(id = R.drawable.ic_connection_error),
+                contentDescription = stringResource(R.string.connection_error)
+            )
+            Spacer(
+                modifier = Modifier.size(
+                    Spaces.spacer.standartWidth,
+                    Spaces.spacer.smallerHeight
+                )
+            )
+            Text(
+                text = errorMessage + stringResource(R.string.singlehero_error_message),
+                fontFamily = interFamily,
+                fontWeight = FontWeight.Bold,
+                fontSize = Sizes.fontSizes.responseError,
+                color = MaterialTheme.colorScheme.onSecondary
+            )
+        }
+        SingleHeroResult(hero = hero, navigateUp = navigateUp)
+    }
+}
