+++ app/src/main/java/com/example/marvel_app_project/ui/pages/singlehero/screens/SingleHeroResult.kt
@@ -1,4 +1,4 @@
-package com.example.marvel_app_project.ui.pages.SingleHero
+package com.example.marvel_app_project.ui.pages.singlehero.screens
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -10,7 +10,6 @@ import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
@@ -19,31 +18,10 @@ import androidx.compose.ui.res.stringResource
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.example.marvel_app_project.R
-import com.example.marvel_app_project.models.HeroUI
+import com.example.marvel_app_project.models.UiLayer.HeroUI
 import com.example.marvel_app_project.ui.components.SingleHeroTextField
 import com.example.marvel_app_project.ui.theme.Sizes
 import com.example.marvel_app_project.ui.theme.Spaces
-
-
-@Composable
-fun SingleHeroScreen(singleHeroUiState: SingleHeroUiState, navigateUp: () -> Unit){
-
-    when(singleHeroUiState){
-        is SingleHeroUiState.Loading -> SingleHeroLoading()
-        is SingleHeroUiState.Success -> SingleHeroResult(
-            hero = singleHeroUiState.singleHeroUIValue,
-            navigateUp = navigateUp
-        )
-        is SingleHeroUiState.Error -> SingleHeroError(
-            errorMessage = singleHeroUiState.errorMessage,
-            hero = singleHeroUiState.reserveSingleHeroUiValue,
-            navigateUp = navigateUp
-        )
-    }
-
-
-
-}
 
 @Composable
 fun SingleHeroResult(hero: HeroUI, navigateUp: () -> Unit){
@@ -79,17 +57,4 @@ fun SingleHeroResult(hero: HeroUI, navigateUp: () -> Unit){
 
         }
     }
-}
-
-@Composable
-fun SingleHeroError(errorMessage: String, hero: HeroUI, navigateUp: () -> Unit){
-    Column {
-        Text(text = "$errorMessage Can't reach a single hero", color = MaterialTheme.colorScheme.onSecondary)
-        SingleHeroResult(hero = hero, navigateUp = navigateUp)
-    }
-}
-
-@Composable
-fun SingleHeroLoading(){
-    Text(text = "Loading...", color = MaterialTheme.colorScheme.onSecondary)
 }
