+++ app/src/main/java/com/example/marvel_app_project/ui/theme/Color.kt
@@ -13,4 +13,4 @@ val Pink40 = Color(0xFF7D5260)
 val Gray = Color(0xFF2A262B)
 val White = Color(0xFFFFFFFF)
 val Black = Color(0xFF000000)
-val BlackShadow = Color(0,0,0,64)
+val Red = Color(0xFF770308)
