+++ app/src/main/java/com/example/marvel_app_project/ui/theme/Shapes.kt
@@ -5,5 +5,6 @@ import androidx.compose.material3.Shapes
 import androidx.compose.ui.unit.dp
 
 internal val Shapes = Shapes(
-    medium = RoundedCornerShape(12.dp)
+    medium = RoundedCornerShape(12.dp),
+    large = RoundedCornerShape(24.dp)
 )
