+++ app/src/main/java/com/example/marvel_app_project/ui/theme/Spaces.kt
@@ -6,6 +6,7 @@ object Spaces {
     val singleHeroColumn = 45.dp
     val shadowElevation = 6.dp
     val chooseHeroColumn = 50.dp
+    val errorColumn = 40.dp
     object singleHeroTextColumn{
         val start = 15.dp
         val bottom = 55.dp
@@ -22,6 +23,7 @@ object Spaces {
     }
     object spacer{
         val standartWidth = 1.dp
+         val smallerHeight = 5.dp
         val standartHeight = 15.dp
         val extendedHeight = 25.dp
         val theMostExtendedHeight = 40.dp
